Cache request token.
The request token does not have to be retrieved and authorized every time the app is run. The point of OAuth is that the request token can be stored and exchanged for an access token at runtime. The tokens are stored in plain-text in the oauth folder. Currently only one user is supported.

However, request token expiration is untested as well as the access token expiration.